requirements.txt
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bcrypt==4.2.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
cryptography==44.0.0
decorator==5.1.1
Deprecated==1.2.15
Django==5.1.5
django-api-forms==0.21.0
django-filter==24.3
django-imap-backend==0.2.3
django-object-checker==1.0.2
django-unfold==0.44.0
fabric==3.2.2
invoke==2.2.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
paramiko==3.5.0
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg==3.2.4
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.2.1
sentry-sdk==2.20.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
